Move Traits management files to component directory + add traits to m2m config
